:sparkles: Move IPresenter to Presenter.IPresenter
